FinanceRoutines

FinanceRoutines

Functions

Index

Functions

FinanceRoutines.import_Funda — Function

import_Funda(wrds_conn; date_range, variables)
import_Funda(; 
    date_range=(Date("1900-01-01"), Date("2030-01-01"), 
    variables::String = "", user="", password="")

Import the funda file from CapitalIQ Compustat on WRDS PostGre server

Arguments

wrds_conn::Connection: An existing PostGreSQL connection to WRDS; creates one if empty

Keywords

date_range::Tuple{Date, Date}: A tuple of dates to select data (limits the download size)

variables::Vector{String}: A vector of String of additional variable to include in the download

user::String: username to log into the WRDS cli; default to ask user for authentication

password::String: password to log into the WRDS cli

Returns

df_funda::DataFrame: DataFrame with compustat funda file

FinanceRoutines.import_MSF — Function

import_MSF(wrds_conn; date_range, variables)
import_MSF(; 
    date_range=(Date("1900-01-01"), Date("2030-01-01"), 
    variables::String = "", user="", password="")

Import the CRSP Monthly Stock File (MSF) from CRSP on WRDS PostGre server

Arguments

wrds_conn::Connection: An existing PostGreSQL connection to WRDS; creates one if empty

Keywords

date_range::Tuple{Date, Date}: A tuple of dates to select data (limits the download size)

variables::Vector{String}: A vector of String of additional variable to include in the download

user::String: username to log into the WRDS cli; default to ask user for authentication

password::String: password to log into the WRDS cli

Returns

df_msf_final::DataFrame: DataFrame with msf crsp file

FinanceRoutines.import_FF3 — Function

import_FF3(frequency::Symbol)

Download and import the Fama-French 3 Factors from Ken French website.

If frequency is unspecified, import the monthly research returns. If frequency is :daily, import the daily research returns.

FinanceRoutines.build_Funda — Function

build_Funda(df_funda::DataFrame; save)

Clean up the compustat funda file download from crsp (see import_Funda)

Arguments

df_funda::DataFrame: A standard dataframe with compustat data (minimum variables are in import_Funda)

Keywords

save::String: Save a gzip version of the data on path $save/funda.csv.gz; Default does not save the data.

Returns

df_funda::DataFrame: DataFrame with compustat funda file "cleaned"

FinanceRoutines.build_MSF — Function

build_MSF(df_msf::DataFrame; save)

Clean up the compustat funda file download from crsp (see import_Funda)

Arguments

df_funda::DataFrame: A standard dataframe with compustat data (minimum variables are in import_Funda)

Keywords

save::String: Save a gzip version of the data on path $save/funda.csv.gz; Default does not save the data.

Returns

df_msf::DataFrame: DataFrame with compustat funda file "cleaned"

FinanceRoutines.open_wrds_pg — Function

Open a postgresql connection on WRDS server

Index

FinanceRoutines.build_Funda

FinanceRoutines.build_MSF

FinanceRoutines.import_FF3

FinanceRoutines.import_Funda

FinanceRoutines.import_MSF

FinanceRoutines.open_wrds_pg