Public Interface

FinanceRoutines Module

FinanceRoutines.build_Funda! — Method

build_Funda!(df_funda::DataFrame; save)

Clean up the compustat funda file download from crsp (see import_Funda)

Arguments

df_funda::DataFrame: A standard dataframe with compustat data (minimum variables are in import_Funda)

Keywords

save::String: Save a gzip version of the data on path $save/funda.csv.gz; Default does not save the data.

clean_cols::Bool: Clean up the columns of the dataframe to be of type Float64; Default is false and leaves the Decimal type intact

Returns

df_funda::DataFrame: DataFrame with compustat funda file "cleaned"

FinanceRoutines.build_MSF! — Method

build_MSF!(df_msf::DataFrame; save)

Clean up the CRSP Monthly Stock File (see import_MSF)

Arguments

df::DataFrame: A standard dataframe with compustat data (minimum variables are in import_Funda)

Keywords

save::String: Save a gzip version of the data on path $save/funda.csv.gz; Default does not save the data.

trim_col::Bool: Only keep a subset of relevant columns in the final dataset

FinanceRoutines.estimate_price_GSW! — Method

estimate_price_GSW!

arguments

- `maturity::Real`: in years

FinanceRoutines.estimate_return_GSW! — Method

estimate_return_GSW!

arguments

- `maturity::Real`: in years
- `frequency::Symbol`: :daily, :monthly, :annual type 
- `type::Symbol`: :log or standard one-period arithmetic return

FinanceRoutines.estimate_yield_GSW! — Method

estimate_yield_GSW!

arguments

- `maturity::Real`: in years

FinanceRoutines.import_FF3 — Method

import_FF3(frequency::Symbol)

Download and import the Fama-French 3 Factors from Ken French website.

If frequency is unspecified, import the monthly research returns. If frequency is :daily, import the daily research returns.

FinanceRoutines.import_Funda — Method

import_Funda(wrds_conn; date_range, variables)
import_Funda(;
    date_range::Tuple{Date, Date} = (Date("1900-01-01"), Dates.today()),
    variables::String = "", user="", password="")

Import the funda file from CapitalIQ Compustat on WRDS Postgres server

Arguments

wrds_conn::Connection: An existing Postgres connection to WRDS; creates one if empty

Keywords

date_range::Tuple{Date, Date}: A tuple of dates to select data (limits the download size)

variables::Vector{String}: A vector of String of additional variable to include in the download

user::String: username to log into the WRDS cli; default to ask user for authentication

password::String: password to log into the WRDS cli

Returns

df_funda::DataFrame: DataFrame with compustat funda file

FinanceRoutines.import_GSW — Method

import_GSW(; date_range)

GSW Curves

arguments

- `date_range::Tuple{Date, Date}`: range for selection of data

FinanceRoutines.import_MSF — Method

import_MSF(wrds_conn; date_range, variables)
import_MSF(; 
    date_range::Tuple{Date, Date} = (Date("1900-01-01"), Dates.today()),
    variables::String = "", user="", password="")

Import the CRSP Monthly Stock File (MSF) from CRSP on WRDS PostGre server

Arguments

wrds_conn::Connection: An existing Postgres connection to WRDS; creates one if empty

Keywords

date_range::Tuple{Date, Date}: A tuple of dates to select data (limits the download size)

variables::Vector{String}: A vector of String of additional variable to include in the download

user::String: username to log into the WRDS cli; default to ask user for authentication

password::String: password to log into the WRDS cli

Returns

df_msf_final::DataFrame: DataFrame with msf crsp file